requirements.txt
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
nltk==3.9.1
matplotlib==3.9.1
streamlit==1.37.1
wordcloud==1.9.3
scikit-learn==1.5.1
